+++ src/provider.ts
@@ -14,8 +14,6 @@ export class Provider implements vscode.TreeDataProvider<Symbol> {
     private _onDidChangeTreeData: vscode.EventEmitter<Symbol | undefined> = new vscode.EventEmitter<Symbol | undefined>();
     readonly onDidChangeTreeData: vscode.Event<Symbol | undefined> = this._onDidChangeTreeData.event;
 
-    // private excludeTestFiles: boolean = true;
-
     constructor(private symbols: Symbol[], private providerType: ProviderType) {
         vscode.commands.executeCommand('setContext', `showGoOutliner${providerType}View`, symbols.length > 0);
         this._onDidChangeTreeData.fire();
@@ -27,11 +25,19 @@ export class Provider implements vscode.TreeDataProvider<Symbol> {
 
     rootItems(): Thenable<Symbol[]> {
         let list = Array<Symbol>();
-        [ItemType.Type, ItemType.Func, ItemType.Var, ItemType.Const].forEach(e => {
-            if (this.countType(e) > 0) {
-                list.push(Symbol.NewRootItem(e));
-            }
-        });
+        
+        if(this.symbols.length === 0) {
+            let note = new Symbol();
+            note.label = "No results.";
+            note.collapsibleState = vscode.TreeItemCollapsibleState.None;
+            list.push(note);
+        } else {
+            [ItemType.Type, ItemType.Func, ItemType.Var, ItemType.Const].forEach(e => {
+                if (this.countType(e) > 0) {
+                    list.push(Symbol.NewRootItem(e));
+                }
+            });
+        }
         return new Promise(resolve => resolve(list));
     }
 
@@ -76,28 +82,6 @@ export class Provider implements vscode.TreeDataProvider<Symbol> {
             return this.rootItems();
         }
         return this.buildItemList(element);
-
-        // if (element) {
-        //     if (element.rootType !== ItemType.None) {
-        //         switch (element.rootType) {
-        //             case ItemType.Type:
-        //                 list = this.symbols.filter(x => x.type === element.rootType && !x.receiver);
-        //                 list.map(x => {
-        //                     x.collapsibleState = this.symbols.some(y => y.receiver === x.label) ? vscode.TreeItemCollapsibleState.Collapsed : vscode.TreeItemCollapsibleState.None;
-        //                 });
-        //                 break;
-        //             case ItemType.Func:
-        //                 list = this.symbols.filter(x => x.type === element.rootType && !x.receiver);
-        //                 break;
-        //             default:
-        //                 list = this.symbols.filter(x => x.type === element.rootType);
-        //                 break;
-        //         }
-        //     } else {
-        //         list = this.symbols.filter(x => element.label === x.receiver);
-        //     }
-        // }
-        // return new Promise(resolve => resolve(list));
     }
 
     private countType(type: ItemType): number {